Move cost types into admin namespace and format ERB files
