+++ modules/costs/app/views/admin/cost_types/_list.html.erb
@@ -46,9 +46,9 @@ See COPYRIGHT and LICENSE files for more details.
         </colgroup>
         <thead>
           <tr>
-            <%= sort_header_tag 'name', :caption => CostType.model_name.human %>
-            <%= sort_header_tag 'unit', :caption => CostType.human_attribute_name(:unit) %>
-            <%= sort_header_tag 'unit_plural', :caption => CostType.human_attribute_name(:unit_plural) %>
+            <%= sort_header_tag "name", caption: CostType.model_name.human %>
+            <%= sort_header_tag "unit", caption: CostType.human_attribute_name(:unit) %>
+            <%= sort_header_tag "unit_plural", caption: CostType.human_attribute_name(:unit_plural) %>
             <th>
               <div class="generic-table--sort-header-outer">
                 <div class="generic-table--sort-header">
@@ -82,15 +82,15 @@ See COPYRIGHT and LICENSE files for more details.
         <tbody>
           <% cost_types.each do |cost_type| %>
           <tr>
-            <%= content_tag :td, link_to(cost_type.name, {:controller => '/cost_types', :action => 'edit', :id => cost_type}) %>
+            <%= content_tag :td, link_to(cost_type.name, { controller: "/admin/cost_types", action: "edit", id: cost_type }) %>
             <%= content_tag :td, cost_type.unit %>
             <%= content_tag :td, cost_type.unit_plural %>
-            <%= content_tag :td, to_currency_with_empty(cost_type.rate_at(@fixed_date)), :class => "currency", :id => "cost_type_#{cost_type.id}_rate"%>
+            <%= content_tag :td, to_currency_with_empty(cost_type.rate_at(@fixed_date)), class: "currency", id: "cost_type_#{cost_type.id}_rate" %>
             <td>
-              <%= form_for cost_type, :url => { :controller => '/cost_types', :action => 'set_rate', :id => cost_type }, method: :put, html: { class: 'inline-label' }  do |f| %>
+              <%= form_for cost_type, url: { controller: "/admin/cost_types", action: "set_rate", id: cost_type }, method: :put, html: { class: "inline-label" }  do |f| %>
                 <label class="hidden-for-sighted" for="<%= "rate_field_#{cost_type.id}" %>"><%= t(:caption_set_rate) %></label>
                 <%= content_tag :input,
-                                '',
+                                "",
                                 value: "",
                                 name: :rate,
                                 size: 7,
@@ -102,25 +102,25 @@ See COPYRIGHT and LICENSE files for more details.
                 </span>
                 <button type="submit"
                         class="button submit_cost_type">
-                  <%= icon_wrapper('icon icon-save', I18n.t(:caption_save_rate)) %>
+                  <%= icon_wrapper("icon icon-save", I18n.t(:caption_save_rate)) %>
                 </button>
               <% end %>
             </td>
-            <%= content_tag :td, cost_type.is_default? ? icon_wrapper('icon icon-checkmark', I18n.t(:general_text_Yes)) : "" %>
+            <%= content_tag :td, cost_type.is_default? ? icon_wrapper("icon icon-checkmark", I18n.t(:general_text_Yes)) : "" %>
             <td class="buttons">
-              <%= form_for cost_type, url:    cost_type_path(cost_type),
+              <%= form_for cost_type, url: admin_cost_type_path(cost_type),
                                       method: :delete,
-                                      html:   { id:    "delete_cost_type_#{cost_type.id}",
-                                                class: 'delete_cost_type',
-                                                title: t(:button_lock) } do |f| %>
+                                      html: { id: "delete_cost_type_#{cost_type.id}",
+                                              class: "delete_cost_type",
+                                              title: t(:button_lock) } do |f| %>
                 <button type="submit"
                         class="button--link submit_cost_type">
-                  <%= icon_wrapper('icon icon-locked', I18n.t(:button_lock)) %>
+                  <%= icon_wrapper("icon icon-locked", I18n.t(:button_lock)) %>
                 </button>
               <% end %>
             </td>
           </tr>
-          <tr style="display:none;" >
+          <tr style="display:none;">
           </tr>
           <% end %>
         </tbody>
