+++ modules/costs/app/views/admin/cost_types/_rate.html.erb
@@ -30,15 +30,14 @@ See COPYRIGHT and LICENSE files for more details.
 <%-
   templated ||= false
   index ||= "INDEX"
-  new_or_existing = rate.new_record? ? 'new' : 'existing'
+  new_or_existing = rate.new_record? ? "new" : "existing"
   id_or_index = rate.new_record? ? index : rate.id
   prefix = "cost_type[#{new_or_existing}_rate_attributes][]"
   id_prefix = "cost_type_#{new_or_existing}_rate_attributes_#{id_or_index}"
 
   @rate = rate
-  error_messages = error_messages_for 'rate'
+  error_messages = error_messages_for "rate"
 -%>
-
 
 <% unless error_messages.blank? %>
   <tr><td colspan="3"><%= error_messages %></td></tr>
@@ -46,17 +45,16 @@ See COPYRIGHT and LICENSE files for more details.
 <%= fields_for prefix, rate, builder: TabularFormBuilder do |rate_form| %>
   <%= content_tag(:tr,
                   id: id_prefix,
-                  data: templated ? { 'row-target': 'true' } : {},
-                  style: templated ? 'display: none;' : ''
-      ) do %>
+                  data: templated ? { "row-target": "true" } : {},
+                  style: templated ? "display: none;" : "") do %>
     <td class="-no-ellipsis">
       <label class="hidden-for-sighted" for="<%= "#{id_prefix}_valid_from" %>"><%= Rate.human_attribute_name(:valid_from) %></label>
       <%= rate_form.date_picker :valid_from,
-                               index: id_or_index,
-                               class: 'date costs-date-picker',
-                               no_field_container: true,
-                               no_label: true,
-                               required: templated ? false : true %>
+                                index: id_or_index,
+                                class: "date costs-date-picker",
+                                no_field_container: true,
+                                no_label: true,
+                                required: templated ? false : true %>
     </td>
     <td class="currency">
       <span class="inline-label">
@@ -79,7 +77,7 @@ See COPYRIGHT and LICENSE files for more details.
       <a href="#"
          class="delete-row-button no-decoration-on-hover"
          data-action="subform#deleteRow">
-        <%= op_icon('icon-context icon-delete', title: t(:button_delete)) %>
+        <%= op_icon("icon-context icon-delete", title: t(:button_delete)) %>
       </a>
     </td>
   <% end %>
