+++ modules/costs/app/views/admin/costs_settings/show.html.erb
@@ -40,7 +40,7 @@ See COPYRIGHT and LICENSE files for more details.
 <%= styled_form_tag(
       admin_costs_settings_path,
       method: :patch,
-      class: 'op-costs-admin-settings'
+      class: "op-costs-admin-settings"
     ) do %>
 
       <section class="form--section">
@@ -63,6 +63,5 @@ See COPYRIGHT and LICENSE files for more details.
         <%- end %>
       </section>
 
-
-  <%= styled_button_tag t(:button_save), class: '-primary -with-icon icon-checkmark' %>
+  <%= styled_button_tag t(:button_save), class: "-primary -with-icon icon-checkmark" %>
 <% end %>
