+++ modules/costs/app/views/hourly_rates/_list_project.html.erb
@@ -28,11 +28,11 @@ See COPYRIGHT and LICENSE files for more details.
 ++#%>
 
 <%
-  rates = @rates unless rates
-  project = @project unless project
+  rates ||= @rates
+  project ||= @project
 
   current_rate = if @all_rates
-                   at_date_in_project_with_ancestors(Date.today, all_rates, project)
+                   at_date_in_project_with_ancestors(Time.zone.today, all_rates, project)
                  else
                    @user.current_rate(project)
                  end
@@ -40,10 +40,10 @@ See COPYRIGHT and LICENSE files for more details.
 <div class="user-rate-history-list">
 <div class="contextual">
   <% if project && User.current.allowed_in_project?({:controller => '/hourly_rates', :action => 'edit'}, project) %>
-    <%= link_to t(:button_update), {:controller => '/hourly_rates', :action => 'edit', :project_id => project, :id => @user}, :class => 'icon icon-edit', :accesskey => accesskey(:edit) %>
+    <%= link_to t(:button_update), { controller: "/hourly_rates", action: "edit", project_id: project, id: @user }, class: "icon icon-edit", accesskey: accesskey(:edit) %>
   <% end %>
 </div>
-<h3><%=h project.name %><%= (" - " + number_to_currency(current_rate.rate)) if current_rate %></h3>
+<h3><%= h project.name %><%= " - #{number_to_currency(current_rate.rate)}" if current_rate %></h3>
 <% if rates.empty? %>
   <%= no_results_box %>
 <% else %>
@@ -91,7 +91,7 @@ See COPYRIGHT and LICENSE files for more details.
           <tr>
             <td style="padding-right: 1em;"><%= rate.valid_from %></td>
             <td class="currency"><%= number_to_currency(rate.rate) %></td>
-            <td><%= rate == current_rate ? icon_wrapper('icon-context icon-checkmark',I18n.t(:general_text_Yes)) : "" %></td>
+            <td><%= rate == current_rate ? icon_wrapper("icon-context icon-checkmark", I18n.t(:general_text_Yes)) : "" %></td>
           </tr>
         <%- end -%>
         </tbody>
