+++ modules/costs/app/views/hourly_rates/_rate.html.erb
@@ -30,7 +30,7 @@ See COPYRIGHT and LICENSE files for more details.
 <%-
   templated ||= false
   index ||= "INDEX"
-  new_or_existing = rate.new_record? ? 'new' : 'existing'
+  new_or_existing = rate.new_record? ? "new" : "existing"
   id_or_index = rate.new_record? ? index : rate.id
   prefix = "user[#{new_or_existing}_rate_attributes][]"
   id_prefix = "user_#{new_or_existing}_rate_attributes_#{id_or_index}"
@@ -40,14 +40,13 @@ See COPYRIGHT and LICENSE files for more details.
   <%= content_tag(:tr,
                   class: "subform-row",
                   id: id_prefix,
-                  data: templated ? { 'row-target': 'true' } : {},
-                  style: templated ? 'display: none;' : ''
-                  ) do %>
+                  data: templated ? { "row-target": "true" } : {},
+                  style: templated ? "display: none;" : "") do %>
     <td>
       <label class="hidden-for-sighted" for="<%= "#{id_prefix}_valid_from" %>"><%= Rate.human_attribute_name(:valid_from) %></label>
       <%= rate_form.date_picker :valid_from,
                                 index: id_or_index,
-                                class: 'date costs-date-picker',
+                                class: "date costs-date-picker",
                                 no_label: true,
                                 no_field_container: true %>
     </td>
@@ -67,7 +66,7 @@ See COPYRIGHT and LICENSE files for more details.
     </td>
     <td class="buttons">
       <a href="#" class="delete-row-button no-decoration-on-hover" data-action="subform#deleteRow">
-        <%= op_icon('icon-context icon-delete', title: t(:button_delete)) %>
+        <%= op_icon("icon-context icon-delete", title: t(:button_delete)) %>
       </a>
     </td>
   <% end %>
