+++ modules/costs/app/views/hourly_rates/show.html.erb
@@ -31,17 +31,17 @@ See COPYRIGHT and LICENSE files for more details.
 
 <%
   breadcrumb_items = if @project
-                       [{href: project_overview_path(@project.identifier), text: @project.name},
-                        {href: projects_budgets_path(@project.identifier), text: I18n.t(:label_budget_plural)}]
+                       [{ href: project_overview_path(@project.identifier), text: @project.name },
+                        { href: projects_budgets_path(@project.identifier), text: I18n.t(:label_budget_plural) }]
                      else
-                       [{ href: home_path, text: organization_name}]
+                       [{ href: home_path, text: organization_name }]
                      end
 %>
 
 <%=
   render Primer::OpenProject::PageHeader.new do |header|
     header.with_title { t(:caption_default_rate_history_for, user: @user.name) }
-    header.with_breadcrumbs([*(breadcrumb_items),
+    header.with_breadcrumbs([*breadcrumb_items,
                              t(:caption_default_rate_history_for, user: @user.name)])
   end
 %>
@@ -49,8 +49,8 @@ See COPYRIGHT and LICENSE files for more details.
 <% if @project %>
   <% html_title t(:caption_rate_history_for, user: @user.name) %>
 
-  <%= render :partial => 'list_project' %>
+  <%= render partial: "list_project" %>
 <% else %>
-  <%= render :partial => 'list_default' %>
-  <%= render :partial => 'list' %>
+  <%= render partial: "list_default" %>
+  <%= render partial: "list" %>
 <% end %>
