+++ modules/costs/app/views/my/timer/_menu.html.erb
@@ -1,1 +1,1 @@
-<%= angular_component_tag 'opce-timer-account-menu' %>
+<%= angular_component_tag "opce-timer-account-menu" %>
