+++ modules/costs/app/views/my/timer/show.html.erb
@@ -3,14 +3,10 @@
     class="op-menu--item-action"
     data-controller="my--timers"
     data-my--timers-start-value="'<%= @timer.created_at %>'"
-    data-application-target="dynamic"
-  >
+    data-application-target="dynamic">
 
     <span
-      data-my--timers-target="elapsedTime"
+      data-my--timers-target="elapsedTime"></span>
 
-    ></span>
-
-   
   </div>
 <% end %>
