+++ modules/costs/app/views/projects/settings/time_entry_activities/_activities.html.erb
@@ -33,7 +33,7 @@ See COPYRIGHT and LICENSE files for more details.
    .each do |activity| %>
   <div class="form--field">
     <%= styled_label_tag "time_entry_activities_project_#{activity.id}_active", activity.name %>
-    <div class="form--field-container" >
+    <div class="form--field-container">
       <%= hidden_field_tag "time_entry_activities_project[][activity_id]",
                            activity.id %>
       <%= styled_check_box_tag "time_entry_activities_project[][active]",
@@ -45,5 +45,5 @@ See COPYRIGHT and LICENSE files for more details.
 <% end %>
 
 <div class="generic-table--action-buttons">
-  <%= styled_button_tag t(:button_save), class: '-primary -with-icon icon-checkmark' %>
+  <%= styled_button_tag t(:button_save), class: "-primary -with-icon icon-checkmark" %>
 </div>
