+++ modules/costs/app/views/projects/settings/time_entry_activities/show.html.erb
@@ -30,7 +30,7 @@ See COPYRIGHT and LICENSE files for more details.
 <%=
   render Primer::OpenProject::PageHeader.new do |header|
     header.with_title { t(:enumeration_activities) }
-    header.with_breadcrumbs([{ href: project_overview_path(@project.id), text: @project.name  },
+    header.with_breadcrumbs([{ href: project_overview_path(@project.id), text: @project.name },
                              { href: project_settings_general_path(@project.id), text: I18n.t(:label_project_settings) },
                              t(:enumeration_activities)])
   end
@@ -40,11 +40,11 @@ See COPYRIGHT and LICENSE files for more details.
   <%= labelled_tabular_form_for @project,
                                 url: project_settings_time_entry_activities_path(@project) do %>
 
-    <%= render partial: '/projects/settings/time_entry_activities/activities',
+    <%= render partial: "/projects/settings/time_entry_activities/activities",
                locals: { project: @project } %>
   <% end %>
 
-  <br/>
+  <br>
 <% else %>
-  <%= no_results_box custom_title: t('projects.settings.activities.no_results_title_text') %>
+  <%= no_results_box custom_title: t("projects.settings.activities.no_results_title_text") %>
 <% end %>
