+++ modules/costs/app/views/users/_rates.html.erb
@@ -32,5 +32,5 @@ See COPYRIGHT and LICENSE files for more details.
   @rates_default = @rates.delete(nil)
 %>
 
-<%= render :partial => 'hourly_rates/list_default' %>
-<%= render :partial => 'hourly_rates/list' %>
+<%= render partial: "hourly_rates/list_default" %>
+<%= render partial: "hourly_rates/list" %>
